Move checkout_package_workdir for packit_service
With checkout_package_workdir inside PackitAPI
also packit service can benefit from it.
